create G1, G', and G2 adjacency matrices
this file is just a test at this point, but it creates G1 randomly
(using David's method in matrix.py). Then it randomly creates an
isomorphism to generate G' (g prime), and then creates the supergraph
G2. It writes these three adjacency matrices to files. Sorry for the
smelly unreadable code, I'll be fixing it today and tomorrow